Appdelegate logic move to UIApplicationShorycutitemExt
